% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_runtoxdrc.R
\name{drc_toxicity}
\alias{drc_toxicity}
\title{Toxicity configuration for \code{runtoxdrc()}}
\usage{
drc_toxicity(
  toxic.lvl = 0.7,
  toxic.type = c("rel", "abs"),
  toxic.direction = c("below", "above", "different"),
  comp.group = 0,
  target.group = NULL
)
}
\arguments{
\item{toxic.lvl}{Numeric. Cutoff point to determine if modelling occurs. Default: 0.7.}

\item{toxic.type}{Character. Indicates if \code{toxic.lvl} is absolute ("abs") or relative ("rel") to some group. Default: "rel".}

\item{toxic.direction}{Character. Indicates if toxicity is a higher ("above"), lower ("below"), or any change ("different"). Default: "below".}

\item{comp.group}{Character/Numeric. Label used for the group values will be compared to if \code{toxic.type} is "rel". Default: 0.}

\item{target.group}{Character/Numeric. Optional. Restrict comparisons to given group(s). Default: NULL.}
}
\value{
A named list containing toxicity determination settings for use in \code{\link[=runtoxdrc]{runtoxdrc()}}.
}
\description{
Defines how toxicity is determined before and during model fitting.
}
\examples{
drc_toxicity()
drc_toxicity(toxic.lvl = 0.5, toxic.direction = "above")

}
\seealso{
\link{config_runtoxdrc}, \code{\link[=runtoxdrc]{runtoxdrc()}}
}
