% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{getmetadata}
\alias{getmetadata}
\title{Generate metadata from a dataframe}
\usage{
getmetadata(dataset, IDcols = NULL, list_obj = NULL, quiet = FALSE)
}
\arguments{
\item{dataset}{A dataframe.}

\item{IDcols}{Optional. Character. Columns given as a vector used in the
identification of data. These columns are preserved in the modified
`dataset` with the first non-blank value. These values should be
identical within observations grouped by `Conc`.}

\item{list_obj}{Optional. List object used for integration with
[runtoxdrc()].}

\item{quiet}{Logical. Indicates if results should be hidden. Defaults
to FALSE.}
}
\value{
A 1 row dataframe of the identifying parameters of an experimental
 replicate. If `list_obj` is provided, returns this within a list as
 `list_obj$metadata`.
}
\description{
Collects identifying or important values from an experimental replicate.
}
