% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtransform.R
\name{logtransform}
\alias{logtransform}
\title{Transform all concerntrations to log 10 scale.}
\usage{
logtransform(dataset, Conc, list_obj = NULL)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of 'dataset' that groups observations.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
A modified dataframe with an additional column, \code{log_Conc}.
}
\description{
Transform all concerntrations to log 10 scale.
}
\examples{
df <- data.frame(x = rep(c(10, 100) each = 3), y = c(3, 5, 4, 20, 40, 60))
logtransform(dataset = df, Conc = x, Response = y)

}
