% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxdrc_preset.R
\name{print.toxdrc_preset}
\alias{print.toxdrc_preset}
\title{Print a toxdrc preset}
\usage{
\method{print}{toxdrc_preset}(x, ...)
}
\arguments{
\item{x}{A `toxdrc_preset`.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a toxdrc preset
}
