% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxdrc-package.R
\docType{package}
\name{toxdrc-package}
\alias{toxdrc}
\alias{toxdrc-package}
\title{toxdrc: Pipeline for Dose-Response Curve Analysis}
\description{
Provides a variety of tools for assessing dose response curves, with an emphasis on toxicity test data. The main feature of this package are modular functions which can be combined through the namesake pipeline, 'runtoxdrc', to automate the analysis for large and complex datasets. This includes optional data preprocessing steps, like outlier detection, solvent effects, blank correction, averaging technical replicates, and much more. Additionally, this pipeline is adaptable to any long form dataset, and does not require specific column or group naming to work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsalole/toxdrc}
  \item Report bugs at \url{https://github.com/jsalole/toxdrc/issues}
}

}
\author{
\strong{Maintainer}: Jack Salole \email{salolej@mcmaster.ca}

}
\keyword{internal}
