% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_runtoxdrc.R
\name{toxdrc_normalization}
\alias{toxdrc_normalization}
\title{Normalization configuration for \code{runtoxdrc()}}
\usage{
toxdrc_normalization(
  blank.correction = FALSE,
  blank.label = "Blank",
  normalize.resp = FALSE,
  relative.label = 0
)

toxdrc_normalization(
  blank.correction = FALSE,
  blank.label = "Blank",
  normalize.resp = FALSE,
  relative.label = 0
)
}
\arguments{
\item{blank.correction}{Logical. Indicates if the response variable should be blank corrected. Default: FALSE.}

\item{blank.label}{Character. Label used for the blank level. Default: "Blank".}

\item{normalize.resp}{Logical. Indicates if response variable should be normalized to a given group. Default: FALSE.}

\item{relative.label}{Character/Numeric. Label or value used for the group values will be normalized to. Default: 0.}
}
\value{
A named list containing normalization configuration for use in \code{\link[=runtoxdrc]{runtoxdrc()}}.

A named list containing normalization configuration for use in \code{\link[=runtoxdrc]{runtoxdrc()}}.
}
\description{
Defines how normalization steps are applied to the response data before model fitting.

Defines how normalization steps are applied to the response data before model fitting.
}
\examples{
toxdrc_normalization()
toxdrc_normalization(blank.correction = TRUE, relative.label = "Control")

toxdrc_normalization()
toxdrc_normalization(blank.correction = TRUE, relative.label = "Control")

}
\seealso{
\link{config_runtoxdrc}, \code{\link[=runtoxdrc]{runtoxdrc()}}

\link{config_runtoxdrc}, \code{\link[=runtoxdrc]{runtoxdrc()}}
}
