% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_runtoxdrc.R
\name{toxdrc_output}
\alias{toxdrc_output}
\title{Output configuration for the runtoxdrc pipeline.}
\usage{
toxdrc_output(
  condense = FALSE,
  sections = c("ID", "effectmeasure", "best_model_name", "effect")
)
}
\arguments{
\item{condense}{Logical. Indicates if the results should be
summarized into a single dataframe. Defaults to TRUE.}

\item{sections}{Character. Columns given as a vector that should be
present in the summary. Defaults to
`c("ID", "effectmeasure", "best_model_name", "effect")`.}
}
\value{
A named list containing output configuration for use
 in [runtoxdrc()].
}
\description{
Defines how [runtoxdrc()] output is returned.
}
\examples{
toxdrc_output()
toxdrc_output(condense = TRUE)

}
\seealso{
[config_runtoxdrc], [runtoxdrc()]
}
