% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageresponse.R
\name{averageresponse}
\alias{averageresponse}
\title{Average response variable}
\usage{
averageresponse(dataset, Conc, Response, IDcols = NULL, list_obj = NULL)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code{dataset} that groups observations (e.g. Conc).}

\item{Response}{Unquoted column name of \code{dataset} with observations (e.g. RFU).}

\item{IDcols}{Optional. Columns given as a vector \code{c("column1", "column2")} used in the identification of data or important metadata. These columns are preserved in the returned dataset with the first value (not NA, NULL, or blank) of these columns within each level of \code{Conc}. Examples of this are metric type (mortality, indicator name), test information (well plate size, test time, test ID). These values should be identical within a testing group.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
A modified dataset with one row for each level of \code{Conc}. The averaged response is given for each level, in addition to the value. If \code{list_obj} provided, returns this within a list. This is primarly for integration wit \code{runtoxdrc} as it saves the \code{pre_average_dataset} to the growing list to track changes. If no \code{list_obj} provided it returns the edited \code{dataset}.
}
\description{
Average the response variable by group or exposure condition.
}
\examples{
averageresponse(
     toxresult,
     Conc = Conc,
     Response = RFU,
     IDcols = c("TestID", "Test_Number", "Dye", "Type", "Replicate")
   )

}
