% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankcorrect.R
\name{blankcorrect}
\alias{blankcorrect}
\title{Blank correct response variable}
\usage{
blankcorrect(
  dataset,
  Conc,
  blank_group = "Blank",
  Response,
  list_obj = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{A dataframe, containing the columns \code{Conc} and \code{Response}.}

\item{Conc}{Bare (unquoted) column name in \code{dataset} that groups the
\code{Response} variable.}

\item{blank_group}{Character. Name of the \code{Conc} level to calculate the
correction value from.}

\item{Response}{Bare (unquoted) column name in \code{dataset} containing
the response variable.}

\item{list_obj}{Optional. List object used for integration with
\code{\link[=runtoxdrc]{runtoxdrc()}}.}

\item{quiet}{Logical. Indicates if results should be hidden. Defaults
to FALSE.}
}
\value{
A modified \code{dataset} with an additional column, \code{c_response}. If
\code{list_obj} is provided, returns this within a list as
\code{list_obj$dataset}, along with statistics of the correction value as
\code{list_obj$blank_stats}.
}
\description{
\code{blankcorrect()} subtracts a calculated correction value
from all responses.
}
\examples{
blankcorrect(
     dataset = toxresult,
     Conc = Conc,
     blank_group = "Blank",
     Response = RFU
   )

}
