% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankcorrect.R
\name{blankcorrect}
\alias{blankcorrect}
\title{Blank correct a response variable.}
\usage{
blankcorrect(
  dataset,
  Conc,
  blank_group = "Blank",
  Response,
  list_obj = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of 'dataset' that groups observations.}

\item{blank_group}{Quoted name of the Blank group, e.g. ("Blank").}

\item{Response}{Unquoted column name of the response variable to be adjusted.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
A modified dataframe with an additional column, \code{c_response}.
}
\description{
Blank correct a response variable.
}
\examples{
blankcorrect(
     toxresult,
     Conc = Conc,
     blank_group = "Blank",
     Response = RFU
   )

}
