% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checktoxicity.R
\name{checktoxicity}
\alias{checktoxicity}
\title{Check for toxic effect}
\usage{
checktoxicity(
  dataset,
  Conc,
  Response,
  effect,
  type = c("rel", "abs"),
  reference_group = "0",
  target_group = NULL,
  list_obj = NULL
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code{dataset} that groups observations (e.g. Conc).}

\item{Response}{Unquoted column name of \code{dataset} with observations (e.g. RFU).}

\item{effect}{Numeric value dictating the point at which observations are flagged as toxic; can be relative or absolute (see \code{type}.)}

\item{type}{Indicates if the effect argument is relative ("rel") to \verb{reference group} or an absolute ("abs") value.}

\item{reference_group}{Quoted name OR value of reference group to compare response to (i.e. "ctl", 0)}

\item{target_group}{Optional. Can be used to limit the compairison to certain groups (highest exposure concentration).}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
The supplied dataset, unchanged, with and a printed message indicating if a test is or is not toxic. If list_obj is supplied, the results of the test  are saved under \code{effect}, where \code{effect = TRUE} indicates an effect that exceeds the threshold.
}
\description{
This test can be used to scan datasets to determine if toxicity has occured. Useful to filter out tests prior to modelling steps.
}
\examples{
df <- data.frame(x = rep(1:2, each = 3), y = c(100, 110, 90, 40, 50, 60))
lt <- list(dataset = df, Group = "B")
checktoxicity(dataset = lt$dataset, Conc = x, Response = y, effect = 85,
type = "abs", list_obj = lt)

}
