% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagCV.R
\name{flagCV}
\alias{flagCV}
\title{Check for groups with high CV}
\usage{
flagCV(dataset, Conc, Response, max_val = 30, update_dataset = FALSE)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of 'dataset' that groups observations to calculate CV in (e.g. Conc).}

\item{Response}{Unquoted column name of 'dataset' with observations to calculate CV in (e.g. RFU).}

\item{max_val}{The maximum CV before a group is flagged.}

\item{update_dataset}{Logical. If TRUE, adds rows to supplied 'dataset' to return. If FALSE, dataset is not changed and results are printed.}
}
\description{
This function calculates the coefficent of variation (CV) of each of the groups, and flags if 'max_val' is exceeded.
}
