% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECx.R
\name{getECx}
\alias{getECx}
\title{Get ECx estimates from model}
\usage{
getECx(
  dataset,
  model,
  EDx = 0.5,
  interval = c("tfls", "fls", "delta", "none"),
  level = 0.95,
  type = c("absolute", "relative"),
  quiet = FALSE,
  EDargs.supplement = list(),
  metadata = NULL,
  list_obj = NULL
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{model}{A drm model, generated by \code{modelcomp()} or \code{drm()}.}

\item{EDx}{The effect measure of interest (i.e. 0.5 for EC50)}

\item{metadata}{Optional. Generated by getmetadata.}

\item{list_obj}{Optional. Used for integration with \code{runtoxdrc}.}

\item{EDargs}{Additional arguments passed to \code{ED()} (e.g., type, interval, level, lower, upper, lref, uref). Should be presented in a list format. See ?ED for more info on arguments.}
}
\value{
If list_obj is not provided, returns estimate. If list object is provided, list_obj$results stores a dataframe with the ECx estimates.
}
\description{
Get ECx estimates from model
}
