% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECx.R
\name{getECx}
\alias{getECx}
\title{Get ECx estimates from model}
\usage{
getECx(
  dataset,
  model,
  EDx = 0.5,
  interval = c("tfls", "fls", "delta", "none"),
  level = 0.95,
  type = c("absolute", "relative"),
  quiet = FALSE,
  EDargs.supplement = list(),
  metadata = NULL,
  list_obj = NULL
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{model}{A drm model, generated by \code{modelcomp()} or \code{drm()}.}

\item{EDx}{The effect measure of interest (i.e. 0.5 for EC50)}

\item{interval}{Character. Method for calculating confidence intervals of EDx. Choices: \code{"tfls"}, \code{"fls"}, \code{"delta"}, \code{"none"}. Default: "tfls".}

\item{level}{Numeric. Confidence level for the interval calculation. Default: 0.95.}

\item{type}{Character. Whether EDx is calculated as \code{"absolute"} or \code{"relative"}. Default: "absolute".}

\item{quiet}{Logical. Whether EDx results should be printed. Default: FALSE.}

\item{EDargs.supplement}{List. Optional user-supplied list of additional or overriding ED arguments. Can include \code{interval}, \code{level}, \code{type}, or other arguments compatible with \code{drc::ED()}.}

\item{metadata}{Optional. Generated by getmetadata.}

\item{list_obj}{Optional. Used for integration with \code{runtoxdrc}.}
}
\value{
If list_obj is not provided, returns estimate. If list object is provided, list_obj$results stores a dataframe with the ECx estimates.
}
\description{
Get ECx estimates from model
}
