% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{getmetadata}
\alias{getmetadata}
\title{Collects identifying values of an expeirmental replicate.}
\usage{
getmetadata(dataset, IDcols, list_obj = NULL)
}
\arguments{
\item{dataset}{A dataframe.}

\item{IDcols}{Optional. Columns given as a vector \code{c("column1", "column2")} used in the identification of data or important metadata. These columns are preserved in the returned dataset with the first value (not NA, NULL, or blank) of these columns within each level of \code{Conc}. Examples of this are metric type (mortality, indicator name), test information (well plate size, test time, test ID). These values should be identical within a testing group.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
A 1 row dataframe of the identifying parameters of an experimental replicate. If list_obj is provided, attaches this to list_obj$metadata.
}
\description{
Collects identifying values of an expeirmental replicate.
}
