% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomp.R
\name{modelcomp}
\alias{modelcomp}
\title{Comparing model fits}
\usage{
modelcomp(
  dataset,
  Conc,
  Response,
  model_list = NULL,
  metric = c("IC", "Res var", "Lack of fit"),
  list_obj = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code{dataset} of independent variable.}

\item{Response}{Unquoted column name of \code{dataset} of dependant variable.}

\item{model_list}{Optional. A list of named model function names (i.e. \code{model_list <- list("LL.4" = LL.4)}).}

\item{metric}{Dictates which model parameter is used to compare model fit.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
A fitetd drm model selected from the . If list_obj is provided, stores this datafram in \code{list_obj$model_df}.
}
\description{
Comparing model fits
}
