% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomp.R
\name{modelcomp}
\alias{modelcomp}
\title{Comparing model fits}
\usage{
modelcomp(
  dataset,
  Conc,
  Response,
  models = NULL,
  models_fx = NULL,
  metric = c("IC", "Res var", "Lack of fit", "no"),
  list_obj = NULL
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code{dataset} of independent variable.}

\item{Response}{Unquoted column name of \code{dataset} of dependant variable.}

\item{models}{Optional. A string of model function names (i.e. \code{"LL.2"}).}

\item{models_fx}{Optional. A list of model functions (i.e \code{LL.2}).}

\item{metric}{Dictates which model parameter is used to compare model fit.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}
}
\value{
A dataframe comparing model fits to the data. If list_obj is provided, stores this datafram in \code{list_obj$model_df}.
}
\description{
Comparing model fits
}
