% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeresponse.R
\name{normalizeresponse}
\alias{normalizeresponse}
\title{Normalize response variables}
\usage{
normalizeresponse(
  dataset,
  Conc,
  reference_group = "0",
  Response,
  list_obj = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code{dataset} that groups observations (e.g. Conc).}

\item{reference_group}{Quoted name OR value of reference group to normalize response to (i.e. "ctl", 0)}

\item{Response}{Unquoted column name of \code{dataset} with observations (e.g. RFU).}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}.}

\item{quiet}{Logical. Whether results should be hidden. Default: FALSE.}
}
\value{
A modified \code{dataset} with a \code{normalized_response} column. If \code{list_obj} provided, updates this within a list. This is primarly for integration wit \code{runtoxdrc} as it adds \code{normalize_response_summary} to the growing list to track changes. If no \code{list_obj} provided, prints the summary and returns the edited \code{dataset}.
}
\description{
Score responses as a percentage of a reference group.
}
\examples{
df <- data.frame(x = rep(1:2, each = 3), y = c(100, 110, 90, 40, 50, 60))
}
