% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctl.R
\name{pctl}
\alias{pctl}
\title{Procedure control check}
\usage{
pctl(
  dataset,
  Conc,
  reference_group = "Control",
  positive_group = 0,
  Response,
  max_diff = 10,
  list_obj = NULL
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code{dataset} with groups to compare (e.g. reference/true control, solvent/positive control)}

\item{reference_group}{Quoted name OR value of reference group, or true control (i.e. "Control", 0)}

\item{positive_group}{Quoted name OR value of procedural group, or positive control (i.e. "ctl+", 0)}

\item{Response}{Unquoted column name of \code{dataset} with observations to compare (e.g. RFU).}

\item{max_diff}{The maximum \% difference between groups before they are flagged.}

\item{list_obj}{Optional existing list object, used for integration with \code{runtoxdrc}. If \code{update_dataset} is set to FALSE, this should be left as NULL.}
}
\value{
A modified dataframe with an additional column, \code{Validity}.
}
\description{
This function evaluates the difference between two groups with the intention of evaluating validity criteria. This can be used for solvent controls, and similar control groups commonly used in toxicity testing, or even to compare more abstract concepts (order of testing, time of day, etc) to validate a method. Careful consideration should be given the max_diff value. The default is set to 10\% difference.
}
\examples{
df <- data.frame(x = rep(1:2, each = 3), y = c(10, 11, 9, 20, 40, 60))
pctl(dataset = pre_df, Conc = x, reference_group = 1, positive_group = 2, Response = y, max_diff = 10)

}
