% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeoutliers.R
\name{removeoutliers}
\alias{removeoutliers}
\title{Remove outliers iteratively using Grubbs' test.}
\usage{
removeoutliers(dataset, Conc, Response, list_obj = NULL)
}
\arguments{
\item{dataset}{A dataframe.}

\item{Conc}{Unquoted column name of \code {dataset} that groups observations to determine outliers, (e.g. Conc).}

\item{Response}{Unquoted column name of \code {dataset} with observations to determine outliers, (e.g. RFU).}

\item{list_obj}{Optional existing list object, used for integration with \code {runtoxdrc}.}
}
\value{
An modified \code{dataset} with outliers removed. If list_obj provided, updates this within a list; primarly for \code {runtoxdrc} and adds \code{$removed_outliers} to the list to track changes. If no list_obj provided, prints the removed rows and returns the edited \code{dataset}.
}
\description{
This function removes statistical outliers from each concentration by iteratively applying Grubbs' test.
}
\examples{

df <- data.frame(x = rep(1:2, each = 3),y = c(3, 5, 7, 3, 4, 30))
remove_outliers(dataset = df, Conc = x, Response = y)
}
