% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtoxdrc.R
\name{runtoxdrc}
\alias{runtoxdrc}
\title{A pipeline for dose response curve modelling and analysis.}
\usage{
runtoxdrc(
  dataset,
  Conc,
  Response,
  IDcols,
  Rep,
  outlier.test = FALSE,
  cv.flag = TRUE,
  cvflag.lvl = 30,
  pctl.test = FALSE,
  pctl.lvl = 10,
  pctl.label = 0,
  ref.label = "Control",
  blank.correction = FALSE,
  blank.label = "Blank",
  normalize.resp = FALSE,
  relative.label = 0,
  avg.resp = TRUE,
  toxic.lvl = 0.7,
  toxic.type = c("rel", "abs"),
  comp.group = 0,
  target.group = NULL,
  non.toxic = c("skip", "continue"),
  model.list = NULL,
  model.metric = c("IC", "Res var", "Lack of fit"),
  EDx = 0.5,
  EDargs = NULL,
  output = c("summary", "full")
)
}
\description{
A pipeline for dose response curve modelling and analysis.
}
