% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxdrc-package.R
\docType{package}
\name{toxdrc-package}
\alias{toxdrc}
\alias{toxdrc-package}
\title{toxdrc: toxdrc}
\description{
toxdrc provides a variety of tools for assessing dose response curves, with an emphasis on toxicity test data. The main feature of this package are modular functions which combine to form pipelines to automate data analysis for large datasets. Dose response curves are fit and point estimation is estimate. Data can also be manipulated during this pipeline (outliers, CV calculation, validity criteria, etc.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsalole/toxdrc}
  \item Report bugs at \url{https://github.com/jsalole/toxdrc/issues}
}

}
\author{
\strong{Maintainer}: Jack Salole \email{salolej@mcmaster.ca}

}
\keyword{internal}
