% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_runtoxdrc.R
\name{toxdrc_output}
\alias{toxdrc_output}
\title{Output configuration for \code{runtoxdrc()}}
\usage{
toxdrc_output(
  condense = FALSE,
  sections = c("ID", "effectmeasure", "best_model_name", "effect")
)
}
\arguments{
\item{condense}{Logical.}

\item{sections}{Character.}
}
\value{
A named list containing output configuration for use in \code{\link[=runtoxdrc]{runtoxdrc()}}.
}
\description{
Defines how the output is returned.
}
\examples{
toxdrc_output()
toxdrc_output(condense = TRUE)

}
\seealso{
\link{config_runtoxdrc}, \code{\link[=runtoxdrc]{runtoxdrc()}}
}
