% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_runtoxdrc.R
\name{toxdrc_qc}
\alias{toxdrc_qc}
\title{Quality control configuration for \code{runtoxdrc()}}
\usage{
toxdrc_qc(
  outlier.test = FALSE,
  cv.flag = TRUE,
  cvflag.lvl = 30,
  pctl.test = FALSE,
  pctl.lvl = 10,
  ref.label = "Control",
  pctl.label = 0,
  avg.resp = TRUE
)
}
\arguments{
\item{outlier.test}{Logical. Indicates if outliers should be identified. Default: FALSE.}

\item{cv.flag}{Logical. Flag CVs of the response variable. Default: TRUE.}

\item{cvflag.lvl}{Numeric. Percent at which CVs are considered high. Default: 30.}

\item{pctl.test}{Logical. Indicates if positive control/solvent effects should be evaluated. Default: FALSE.}

\item{pctl.lvl}{Numeric. Percent difference in response \code{ref.label} and \code{pctl.label} at which tests are flagged. Default: 10.}

\item{ref.label}{Character. Label used for the true control level. Default: "Control".}

\item{pctl.label}{Character. Label used for the positive control level. Default: 0.}

\item{avg.resp}{Logical. Indicates if responses should be averaged at each concentration. Default: TRUE.}
}
\value{
A named list containing the quality control configuration for use in \code{\link[=runtoxdrc]{runtoxdrc()}}.
}
\description{
Defines how quality control is applied to the response data before model fitting.
}
\examples{
toxdrc_qc()
toxdrc_qc(outlier.test = TRUE, cvflag.lvl = 20)

}
\seealso{
\link{config_runtoxdrc}, \code{\link[=runtoxdrc]{runtoxdrc()}}
}
